requirements.txt
torch==2.1.0
transformers==4.35.0
wandb==0.15.12
numpy==1.24.3
pandas==2.1.1
tqdm==4.66.1
python-dotenv==1.0.0
pydantic==2.4.2
peft==0.6.0
urllib3==1.26.18
requests==2.31.0 
